% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aea_ads.R
\name{aea_ads}
\alias{aea_ads}
\title{Title Approximate Exchange Algorithm (AEA) with Adaptive Direction Sampling (ADS)}
\usage{
aea_ads(
  formula,
  y_t = y_t,
  L.x_F = L.x_F,
  L.x_P = L.x_P,
  L.z_F = L.z_F,
  L.z_P = L.z_P,
  sy = sy,
  iters = 50000,
  burn.in = 100,
  ergm_sim_iters = 1000,
  mu.prior_F = NULL,
  Sigma.prior_F = NULL,
  mu.prior_P = NULL,
  Sigma.prior_P = NULL,
  n_chains = 10,
  gamma_F = 0.5,
  gamma_P = 0.5,
  sigma.prop_F = 0.025,
  sigma.prop_P = 0.025
)
}
\arguments{
\item{formula}{2-layer ERGM model formula}

\item{y_t}{list of observed multi-layer networks}

\item{L.x_F}{list of observed x_F networks}

\item{L.x_P}{list of observed x_P networks}

\item{L.z_F}{list of observed z_F networks}

\item{L.z_P}{list of observed z_P networks}

\item{sy}{observed statistics, i.e., list(sz$F_obs, sz$P_obs), include the L(~edges, ~(active))}

\item{iters}{overall number of iterations for each MCMC chain}

\item{burn.in}{number of burnin iterations for each MCMC chain}

\item{ergm_sim_iters}{number of iterations for ERGM simulation}

\item{mu.prior_F}{prior mean for zeta_F}

\item{Sigma.prior_F}{prior covariance for zeta_F}

\item{mu.prior_P}{prior mean for zeta_P}

\item{Sigma.prior_P}{prior covariance for zeta_P}

\item{n_chains}{number of MCMC chains}

\item{gamma_F}{ADS move factor for the F process}

\item{gamma_P}{ADS move factor for the P process}

\item{sigma.prop_F}{diagonal entry for the multivariate Normal proposal for the F process}

\item{sigma.prop_P}{diagonal entry for the multivariate Normal proposal for the P process}
}
\value{
posterior samples of theta_F and theta_P
}
\description{
Title Approximate Exchange Algorithm (AEA) with Adaptive Direction Sampling (ADS)
}
