% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aea_amh.R
\name{aea_amh}
\alias{aea_amh}
\title{Title Approximate Exchange Algorithm (AEA) with Adaptive Metropolis Hastings (AMH)}
\usage{
aea_amh(
  formula,
  y_t = y_t,
  L.x_F = L.x_F,
  L.x_P = L.x_P,
  L.z_F = L.z_F,
  L.z_P = L.z_P,
  sy = sy,
  iters = 50000,
  burn.in = 100,
  ergm_sim_iters = 1000,
  mu.prior_F = NULL,
  Sigma.prior_F = NULL,
  mu.prior_P = NULL,
  Sigma.prior_P = NULL,
  n_chains = 1,
  init_s_F = 0.025,
  init_s_P = 0.025,
  adapt_start = 100,
  adapt_interval = 50
)
}
\arguments{
\item{formula}{2-layer ERGM model formula}

\item{y_t}{list of observed multi-layer networks}

\item{L.x_F}{list of observed x_F networks}

\item{L.x_P}{list of observed x_P networks}

\item{L.z_F}{list of observed z_F networks}

\item{L.z_P}{list of observed z_P networks}

\item{sy}{observed statistics, i.e., list(sz$F_obs, sz$P_obs), include the L(~edges, ~(active))}

\item{iters}{overall number of iterations for each MCMC chain}

\item{burn.in}{number of burnin iterations for each MCMC chain}

\item{ergm_sim_iters}{number of iterations for ERGM simulation}

\item{mu.prior_F}{prior mean for zeta_F}

\item{Sigma.prior_F}{prior covariance for zeta_F}

\item{mu.prior_P}{prior mean for zeta_P}

\item{Sigma.prior_P}{prior covariance for zeta_P}

\item{n_chains}{number of MCMC chains, default = 1}

\item{init_s_F}{initial scaling factor for proposal covariance for the F process}

\item{init_s_P}{initial scaling factor for proposal covariance for the P process}

\item{adapt_start}{when to start adaptation}

\item{adapt_interval}{how often to update covariance}
}
\value{
posterior samples of theta_F and theta_P
}
\description{
Title Approximate Exchange Algorithm (AEA) with Adaptive Metropolis Hastings (AMH)
}
