% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senators70_am.R
\docType{data}
\name{senators70_am}
\alias{senators70_am}
\title{Adjacency matrices of 70 Senators from 99 to 102 U.S. Congress}
\format{
## `senators70_am`
A list of 4 70 x 70 signed adjacency matrices with attributes about ID, Senator's name, State and Party.
}
\source{
<https://figshare.com/articles/dataset/A_Sign_of_the_Times/8096429?file=20758701>
}
\usage{
senators70_am
}
\description{
A subset of data from Neal, Z. P. (2014). The backbone of bipartite projections: Inferring relationships from co-authorship, co-sponsorship, co-attendance and other co-behaviors.
Social Networks, 39, 84-97, and Neal, Z. P. (2020). A sign of the times? Weak and strong polarization in the US Congress,
1973–2016. Social Networks, 60, 103-112.
}
\keyword{datasets}
