% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stats.R
\name{simulate_stats}
\alias{simulate_stats}
\title{Title Simulate Network Statistics from a Separable 2-layer TERGM}
\usage{
simulate_stats(
  formula,
  y_t = y_t,
  L.x_F = L.x_F,
  L.x_P = L.x_P,
  L.z_F = L.z_F,
  L.z_P = L.z_P,
  times = NULL,
  model.dim = NULL,
  coef_F = NULL,
  coef_P = NULL,
  ergm_sim_iters = 1000,
  network = FALSE,
  network.stats = FALSE,
  ...
)
}
\arguments{
\item{formula}{2-layer ERGM model formula}

\item{y_t}{list of multi-layer networks}

\item{L.x_F}{list of observed x_F networks}

\item{L.x_P}{list of observed x_P networks}

\item{L.z_F}{initial z_F state (multi-layer network)}

\item{L.z_P}{initial z_P state (multi-layer network)}

\item{times}{number of time points, including t_0}

\item{model.dim}{model dimension, including constraint: L(~edges, ~(active))}

\item{coef_F}{parameter zeta_F without constraint: (Inf,...}

\item{coef_P}{parameter zeta_P without constraint: (Inf,...}

\item{ergm_sim_iters}{number of iterations for ERGM simulation}

\item{network}{logical, if TRUE saves the simulated networks z_0:T}

\item{network.stats}{logical, if TRUE saves the simulated network statistics of z_0:T}

\item{...}{additional arguments}
}
\value{
list of network statistics and network objects
}
\description{
Title Simulate Network Statistics from a Separable 2-layer TERGM
}
